refactor: migrate init to src dir
