+++ pages/showcase/forms/AddressFormExample.tsx
@@ -5,11 +5,11 @@ import { BrothTypography } from '@/components/typography/BrothTypography';
 import { Box, Container, Grid, Stack } from '@mui/material';
 import DeleteIcon from '@mui/icons-material/Delete';
 import { useForm } from 'react-hook-form';
-import yup from 'init/yup';
+import yup from '@/init/yup';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { useSnackbar } from 'hooks/useSnackbar';
 import BrothSnackbar from '@/components/snackbar/BrothSnackbar';
-import { YubinBangoCore } from 'init/yubinbango-core';
+import { YubinBangoCore } from '@/init/yubinbango-core';
 
 interface formTypes {
   zip_code: string;
