+++ src/init/swagger/core/BaseHttpRequest.ts
