+++ src/init/swagger/core/CancelablePromise.ts
