+++ src/init/swagger/models/CategoryEnum.ts
