+++ src/init/swagger/models/PatchedUser.ts
