+++ src/init/swagger/models/ProviderAuth.ts
