+++ src/init/swagger/services/LogoutService.ts
