+++ src/init/yubinbango-core.ts
